+++ src/components/api-response.js
@@ -2,7 +2,8 @@ import { LitElement, html, css } from 'lit-element';
 import { marked } from 'marked';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { schemaInObjectNotation, generateExample } from '../utils/schema-utils';
-import FontStyles from '../styles/font-styles';
+import { getI18nText } from '../languages';
+import FontStyles from '../styles/font-styles.js';
 import FlexStyles from '../styles/flex-styles';
 import TableStyles from '../styles/table-styles';
 import InputStyles from '../styles/input-styles';
@@ -25,6 +26,7 @@ export default class ApiResponse extends LitElement {
       callback: { type: String },
       responses: { type: Object },
       parser: { type: Object },
+      includeNulls: { type: Boolean, attribute: 'display-nulls', converter(value) { return value === 'true'; } },
       schemaStyle: { type: String, attribute: 'schema-style' },
       renderStyle: { type: String, attribute: 'render-style' },
       selectedStatus: { type: String, attribute: 'selected-status' },
@@ -97,7 +99,7 @@ export default class ApiResponse extends LitElement {
     return html`
     <div class="col regular-font response-panel ${this.renderStyle}-mode">
       <div class=" ${this.callback === 'true' ? 'tiny-title' : 'req-res-title'} "> 
-        ${this.callback === 'true' ? 'CALLBACK RESPONSE' : 'RESPONSE'}
+        ${this.callback === 'true' ? 'CALLBACK RESPONSE' : getI18nText('operations.response')}
       </div>
       <div>
         ${this.responseTemplate()}
@@ -125,7 +127,7 @@ export default class ApiResponse extends LitElement {
           this.selectedMimeType = mimeResp;
         }
         // Generate Schema
-        const schemaTree = schemaInObjectNotation(mimeRespObj.schema, {});
+        const schemaTree = schemaInObjectNotation(mimeRespObj.schema, { includeNulls: this.includeNulls });
         // Generate Example
         const respExamples = generateExample(
           (mimeRespObj.examples || ''),
@@ -169,7 +171,7 @@ export default class ApiResponse extends LitElement {
               part="btn--resp ${this.selectedStatus === respStatus ? 'btn-fill--resp' : 'btn-outline--resp'} btn-response-status"
               style='margin: 8px 4px 0 0'
             > 
-              ${respStatus} 
+              ${respStatus}
             </button>`
           }`)
         }
@@ -189,8 +191,8 @@ export default class ApiResponse extends LitElement {
             : html`  
               <div class="tab-panel col">
                 <div class="tab-buttons row" @click="${(e) => { if (e.target.tagName.toLowerCase() === 'button') { this.activeSchemaTab = e.target.dataset.tab; } }}" >
-                  <button class="tab-btn ${this.activeSchemaTab === 'model' ? 'active' : ''}" data-tab='model'>MODEL</button>
-                  <button class="tab-btn ${this.activeSchemaTab === 'body' ? 'active' : ''}" data-tab='body'>EXAMPLE</button>
+                  <button class="tab-btn ${this.activeSchemaTab === 'model' ? 'active' : ''}" data-tab='model'>${getI18nText('operations.model')}</button>
+                  <button class="tab-btn ${this.activeSchemaTab === 'body' ? 'active' : ''}" data-tab='body'>${getI18nText('operations.example')}</button>
                   <div style="flex:1"></div>
                   ${Object.keys(this.mimeResponsesForEachStatus[status]).length === 1
                     ? html`<span class='small-font-size gray-text' style='align-self:center; margin-top:8px;'> ${Object.keys(this.mimeResponsesForEachStatus[status])[0]} </span>`
@@ -214,7 +216,7 @@ export default class ApiResponse extends LitElement {
 
   responseHeaderListTemplate(respHeaders) {
     return html`
-      <div style="padding:16px 0 8px 0" class="resp-headers small-font-size bold-text">RESPONSE HEADERS</div> 
+      <div style="padding:16px 0 8px 0" class="resp-headers small-font-size bold-text">${getI18nText('operations.response-headers')}</div> 
       <table role="presentation" style="border-collapse: collapse; margin-bottom:16px; border:1px solid var(--border-color); border-radius: var(--border-radius)" class="small-font-size mono-font">
         ${respHeaders.map((v) => html`
           <tr>