+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 ## 1.0 ##
 * By default uses the `info.title` property of the open api specification for the `overview` section name in the navbar. Falls back to the default when not specified.
 
+## 0.10 ##
+* Internationalization support with languages `en` and `fr`
+
 ## 0.9 ##
 * OpenAPI 3.1 support using `APIDevTools/json-schema-ref-parser`
 * Enable searching for schemas using the search filter.
@@ -17,6 +20,9 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * Fix issues with allOf/oneOf with partially completed data.
 * Sanitize paths coming from the spec that contain invalid characters.
 * Improve display of array types and number/string formats in parameters and models
+* Fix fetch request options to not require unnecessary extra level in setting properties.
+* Fix support for `const` in json schema
+* Fix support for `null` types and add property `display-nulls` to support forcing the display of them in models.
 
 ## 0.8 ##
 * Removing the `RESET` button as it is confusing for users to see. It only repopulated the defaults, and to do that, you can easily switch tabs and come back.